Simplify Finnhub Security Search Provider following review comments
Combined code from FinnhubSymbolSearch into FinnhubSearchProvider as
this had just been copied from the YahooSearchProvider, which does two
types of searches.

Do not search for duplicates in FinnhubSearchProvider results as that is
already doen by Finnhub. This code had just been copied from the
YahooSearchProvider, which does two types of searches.

Factory now gets SecuritySearchProvider base on the implementation class
instead of ID. This is the same way that the DividendFeed is found.
